modificacao da multihal

+++ multihal/CompleteSensor.h
@@ -106,7 +106,7 @@ class AirQualitySensor : public OnChangeSensor {
         std::vector<Event> readEvents() override;
 
     protected:
-        float sensorRead = 0;
+        int sensorRead = 0;
 
     static int64_t timevalToNano(timeval const& t) {
         return t.tv_sec*1000000000LL + t.tv_usec*1000;
